Made all fields in Hotel/Floor/Room private.
One of the principles of OOP is encapsulation, i.e. that the
implementation of a class is hidden. The primary benefit of this is that
it allows us to change the implementation of a class without the
potential of breaking anything that is currently using that class.

In this example the hotel was exposing the floors array, and the floors
were exposing their rooms array. This was then being used by the main
class to set the clean state of rooms. By making these fields private
and only adding methods to make the rooms clean/dirty it gives us the
flexibility to change things.

Say, for example, that our hotel owner is superstitious and doesn't have
rooms using the number 13. It would be quite a hassle to handle this if
the floor and room arrays are being accessed directly. Now that the
fields are private we could start by simply adding a check in the Floor
class to handle this without needing to change any of the Driver or
Hotel code. On a more technical level this also allows us to change the
way the rooms are stored, we could change to using an ArrayList or
Rooms, or a Map with the room numbers as keys and the Driver class would
never know anything had changed.