Pulled user interaction code up into Driver.java.
There were two reasons I did this. The first is that it allows all the
user interaction code to live in Driver.

The second is that it makes the Hotel class more general. Now that a
Hotel can be constructed to a given specification it could be used
elsewhere in situations where there isn't a user to give it input.

I did consider using a builder pattern for constructing Hotels but since
you said this is for giving instruction to CS students I figured it was
best to do it this way for now. If Hotel construction was more complex
(e.g. allowing non-sequential floor/room numbers, setting up rooms in
different ways, etc) then I'd definitely recommend using a different
construction method.

As a bonus, having all the user interaction code in the Driver solves
the problem of not being able to use (close) multiple Scanners.