Made fields immutable where appropriate.
This is good practice and something I'd recommend everyone get in the
habit of doing, in my opinion there's no such thing as overuse of the
final keyword!

When you know fields aren't going to change - mark them as final. It may
not seem like much but it has some very nice benefits. It you mark a
field as final it means you ('you' here could be you, or anyone else
working with the code) can't accidently change the value of something
you expected to be constant. In this case we would never expect the
number of floors or rooms or the floor number to change once it's been
constructed so I've marked them as final. Of course, if in future we
wanted to allow our Hotel owner to build more floors we could relax this
and remove the final keyword where appropriate to allow this since we
would no longer have any expectation that the number was going to be
fixed once it was set.

There are technical benefits to this as well, final fields are
inherently thread-safe.