Changed floor and room input validation.
I felt it made more sense, in the spirit of encapsulation, to create
methods in Hotel for querying if a given floor or room exists. The
previous validation method was to query the number of floors and the
number of rooms on that floor and check that the input values were
within these limits. What I've done here is create methods on Hotel to
check if a floor or room exists, which is really what we want to know.
This has the benefits of making the Driver code read a bit more nicely
and improves the encapsulation of the Hotel classes. We can now change
the implementation of the hasFloor/hasRoom and, for example, say that we
don't have a room 13 or generally change the way rooms were numbered.